refactor: move gossip metrics helpers into dedicated modules
